Enable render

+++ deploy/deploy.sh
@@ -3,11 +3,15 @@ set -e # Exit with nonzero exit code if anything fails
 
 # Update generated data file
 cd awesome-public-datasets
+if [ ! -e ../deploy/index.rst ]; then
+    echo "ERROR:Can not find rendered meta file. Please run render.py first ahead"
+    exit 1
+fi
 cp ../deploy/index.rst README.rst
 
 # If there are no changes to the compiled out (e.g. this is a README update) then just bail.
 if git diff --quiet; then
-    echo "No changes to the output on this push; exiting."
+    echo "INFO:No changes to the output on this push; exiting."
     exit 0
 fi
